fix: add swiftLint to Scripts
